feat: Getting started with improving portfolio model class
